requirements.txt
os-service-types
requests
pyyaml